+++ app/src/main/java/com/umc/ttoklip/presentation/hometown/communication/CommunicationActivity.kt
@@ -1,4 +1,4 @@
-package com.umc.ttoklip.presentation.hometown
+package com.umc.ttoklip.presentation.hometown.communication
 
 import android.content.Intent
 import androidx.activity.viewModels
@@ -14,6 +14,8 @@ import com.umc.ttoklip.presentation.alarm.AlarmActivity
 import com.umc.ttoklip.presentation.base.BaseActivity
 import com.umc.ttoklip.presentation.hometown.adapter.CommunicationAdapter
 import com.umc.ttoklip.presentation.hometown.adapter.OnItemClickListener
+import com.umc.ttoklip.presentation.hometown.communication.read.ReadCommunicationActivity
+import com.umc.ttoklip.presentation.hometown.communication.write.WriteCommunicationActivity
 import com.umc.ttoklip.presentation.mypage.SortSpinnerAdapter
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
