+++ app/src/main/java/com/umc/ttoklip/presentation/hometown/communication/CommunicationViewModel.kt
@@ -1,4 +1,4 @@
-package com.umc.ttoklip.presentation.hometown
+package com.umc.ttoklip.presentation.hometown.communication
 
 import com.umc.ttoklip.data.model.town.Communities
 import kotlinx.coroutines.flow.StateFlow
