+++ app/src/main/java/com/umc/ttoklip/presentation/hometown/communication/read/ReadCommunicationViewModel.kt
@@ -1,13 +1,10 @@
-package com.umc.ttoklip.presentation.hometown
+package com.umc.ttoklip.presentation.hometown.communication.read
 
-import com.umc.ttoklip.data.model.honeytip.CommentResponse
-import com.umc.ttoklip.data.model.town.CreateCommentRequest
 import com.umc.ttoklip.data.model.town.ReportRequest
 import com.umc.ttoklip.data.model.town.ViewCommunicationResponse
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
 
 interface ReadCommunicationViewModel {
     val postId: StateFlow<Long>
