+++ app/src/main/java/com/umc/ttoklip/presentation/hometown/communication/write/WriteCommunicationViewModelImpl.kt
@@ -1,4 +1,4 @@
-package com.umc.ttoklip.presentation.hometown
+package com.umc.ttoklip.presentation.hometown.communication.write
 
 import android.content.Context
 import android.util.Log
