+++ app/src/main/java/com/umc/ttoklip/presentation/hometown/together/write/WriteTogetherViewModel.kt
@@ -1,4 +1,4 @@
-package com.umc.ttoklip.presentation.hometown
+package com.umc.ttoklip.presentation.hometown.together.write
 
 import com.umc.ttoklip.presentation.honeytip.adapter.Image
 import kotlinx.coroutines.flow.StateFlow
