+++ app/src/main/java/com/umc/ttoklip/presentation/honeytip/write/WriteHoneyTipActivity.kt
@@ -55,7 +55,7 @@ class WriteHoneyTipActivity :
     BaseActivity<ActivityWriteHoneyTipBinding>(R.layout.activity_write_honey_tip),
     OnImageClickListener {
     private val imageAdapter: ImageRVA by lazy {
-        ImageRVA(this)
+        ImageRVA(this, this)
     }
     private val board: String by lazy {
         Log.d("board", intent.getStringExtra(BOARD)!!)
@@ -128,6 +128,7 @@ class WriteHoneyTipActivity :
         if (!isEdit){
             return
         } else {
+            viewModel.isEdit.value = true
             viewModel.isWriteDoneBtnEnable.value = true
             val editHoneyTip = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
                 intent.getSerializableExtra("honeyTip", EditHoneyTip::class.java)
@@ -135,8 +136,6 @@ class WriteHoneyTipActivity :
                 intent.getSerializableExtra("honeyTip") as EditHoneyTip
             }
             with(binding) {
-                writeDoneBtn.isEnabled = true
-                writeDoneBtn.text = "수정완료"
                 titleEt.setText(editHoneyTip?.title)
                 bodyEt.setText(editHoneyTip?.content)
                 inputUrlEt.setText(editHoneyTip?.url)
@@ -147,6 +146,7 @@ class WriteHoneyTipActivity :
             val images = editHoneyTip?.image?.toList()
             Log.d("edit images", images.toString())
             convertURLtoURI(images)
+            imageAdapter.submitList(images?.map { Image(Uri.EMPTY ,it) })
             postId = editHoneyTip?.postId ?: 0
             category = editHoneyTip?.category?:""
             binding.tabLayout.selectTab(binding.tabLayout.getTabAt(stringToNum(category)))
@@ -156,7 +156,6 @@ class WriteHoneyTipActivity :
     private fun convertURLtoURI(photos: List<String>?) {
         val uris = mutableListOf<Uri>()
         photos?.forEach {
-            Log.d("uris", uris.toString())
             Glide.with(this).asBitmap().load(it)
                 .into(object : CustomTarget<Bitmap>() {
                     override fun onResourceReady(
@@ -166,7 +165,8 @@ class WriteHoneyTipActivity :
                         Log.d("bittmap", resource.toString())
                         uris.add(getImageUri(this@WriteHoneyTipActivity, resource))
                         if (uris.size == photos.size) {
-                            imageAdapter.submitList(uris.map { it -> Image(it) })
+                            Log.d("uris", uris.toString())
+                            //imageAdapter.submitList(uris.map { it -> Image(it, "") })
                             editImage.addAll(uris)
                         }
                     }
@@ -256,8 +256,11 @@ class WriteHoneyTipActivity :
 
     private fun writeDone() {
         binding.writeDoneBtn.setOnClickListener {
-            val images =
-                imageAdapter.currentList.filterIsInstance<Image>().map { it.uri }.toList()
+            val beforeEditImages =
+                imageAdapter.currentList.filterIsInstance<Image>().map { it.url }.toList()
+            //convertURLtoURI(beforeEditImages)
+            val images = editImage + imageAdapter.currentList.filterIsInstance<Image>().map { it.uri }.filter { it != Uri.EMPTY }.toList()
+            Log.d("write done", images.size.toString())
             val imageParts = WriteHoneyTipUtil(this).convertUriListToMultiBody(images)
             imageParts.forEach {
                 Log.d("용량", "${it.body.contentLength().toDouble() / (1024 * 1024)}")
@@ -269,13 +272,13 @@ class WriteHoneyTipActivity :
 
             val title = binding.titleEt.text.toString()
             val content = binding.bodyEt.text.toString()
-            val category = category.toString()
-            Log.d("write done category", category)
+            val category = category
             val url = binding.inputUrlEt.text.toString()
 
             if(isEdit){
                 Log.d("it Edit", isEdit.toString())
                 viewModel.editHoneyTip(postId, title, content, category, imageParts, url)
+                Log.d("edit image imagepart", imageParts.contentToString())
                 editImage.forEach{
                     val delete = deleteImage(this@WriteHoneyTipActivity, it)
                     Log.d("delete", delete.toString())
@@ -287,19 +290,6 @@ class WriteHoneyTipActivity :
                     viewModel.createQuestion(title, content, category, imageParts)
                 }
             }
-        }
-    }
-
-    fun getFileSize(filePath: String): Long {
-        val file = File(filePath)
-
-        // 파일이 존재하는지 확인
-        if (file.exists()) {
-            // 파일 크기를 바이트 단위로 반환
-            return file.length()
-        } else {
-            // 파일이 존재하지 않으면 -1 또는 다른 값으로 처리
-            return -1
         }
     }
 
@@ -397,7 +387,7 @@ class WriteHoneyTipActivity :
             applicationContext.contentResolver.takePersistableUriPermission(it, flag)
         }
 
-        val images = uriList.map { Image(it) }
+        val images = uriList.map { Image(it, "") }
         val updatedImages = imageAdapter.currentList.toMutableList().apply { addAll(images) }
         imageAdapter.submitList(updatedImages)
     }
