+++ app/src/main/java/com/umc/ttoklip/presentation/mypage/MyHoneyTipActivity.kt
@@ -12,7 +12,7 @@ import com.umc.ttoklip.R
 import com.umc.ttoklip.data.model.honeytip.HoneyTipMain
 import com.umc.ttoklip.databinding.ActivityMyHoneyTipBinding
 import com.umc.ttoklip.presentation.base.BaseActivity
-import com.umc.ttoklip.presentation.hometown.ReadCommunicationActivity
+import com.umc.ttoklip.presentation.hometown.communication.read.ReadCommunicationActivity
 import com.umc.ttoklip.presentation.honeytip.adapter.HoneyTipListRVA
 import com.umc.ttoklip.presentation.honeytip.adapter.OnItemClickListener
 import com.umc.ttoklip.presentation.honeytip.read.ReadHoneyTipActivity
