+++ app/src/main/java/com/umc/ttoklip/presentation/search2/fragment/SearchTownFragment.kt
@@ -11,8 +11,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.umc.ttoklip.R
 import com.umc.ttoklip.databinding.FragmentItemSearchBinding
 import com.umc.ttoklip.presentation.base.BaseFragment
-import com.umc.ttoklip.presentation.hometown.ReadCommunicationActivity
-import com.umc.ttoklip.presentation.honeytip.read.ReadHoneyTipActivity
+import com.umc.ttoklip.presentation.hometown.communication.read.ReadCommunicationActivity
 import com.umc.ttoklip.presentation.mypage.SortSpinnerAdapter
 import com.umc.ttoklip.presentation.news.detail.ArticleActivity
 import com.umc.ttoklip.presentation.search.adapter.SearchRVA
