+++ app/src/main/res/layout/activity_write_communication.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="com.umc.ttoklip.presentation.hometown.WriteCommunicationViewModelImpl" />
+            type="com.umc.ttoklip.presentation.hometown.communication.write.WriteCommunicationViewModelImpl" />
 
     </data>
