+++ app/src/main/res/layout/activity_write_honey_tip.xml
@@ -4,6 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <data>
+        <import type="android.view.View"/>
+
         <variable
             name="viewModel"
             type="com.umc.ttoklip.presentation.honeytip.write.WriteHoneyTipViewModel" />
@@ -40,7 +42,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginEnd="24dp"
-            android:text="작성완료"
+            android:text='@{viewModel.isEdit ? "수정완료":"작성완료"}'
             android:enabled="@{viewModel.isWriteDoneBtnEnable}"
             android:textAppearance="@style/TextAppearance.App.12sp_400"
             android:textColor="@color/write_btn_selector"
